Support separate plugin path (if it's installed)
When baking, CakePHP's choice of path to bake into depends on whether plugin
being developed was added to CakePHP application in which we're baking.

If the plugin has a working prototype and was added to CakePHP via
composer then bake will put the files directly into it. Bake.sh needs to
be aware of both scenarios.